+++ src/types.ts
@@ -1,10 +1,21 @@
-export interface WithPersistentOptions<Value = any, Serialized = Value> {
+import { Store } from 'effector';
+
+export interface WithPersistentOptions<
+  State = any,
+  Value = State,
+  Serialized = Value
+> {
   /**
    * Debounce subsequent store updates and flush only after latest change.
    * If set to `undefined` (default), no debounce will be used, so every store
    * update will be flushed to driver.
    */
   flushDelay?: number;
+  /**
+   * Alternative target which will receive initial state read from driver on
+   * initialization.
+   */
+  wakeUp?: Store<State> | ((state: State) => void);
   /**
    * Serialization before writing data to driver.
    */